represent orgId as uint32 instead of int
should result in memory savings in the index
when loading for cassandra, if <0 -> assign OrgIdPublic

+++ cmd/mt-index-migrate/main.go
@@ -156,7 +156,7 @@ func getDefs(session *gocql.Session, defsChan chan *schema.MetricDefinition) {
 		}
 		mdef := schema.MetricDefinition{
 			Id:         mkey,
-			OrgId:      orgId,
+			OrgId:      uint32(orgId),
 			Partition:  partition,
 			Name:       name,
 			Interval:   interval,
